Run translation validation on tracing error.
This PR wraps `InstructionTranslator` run with a try-catch block so as to run the
translation validation (TV) if it ends up raising an error.

In this context, we run TV so as to catch simplification errors. These may turn
`ShapeEnv.divisible` and `ShapeEnv.replacements` incorrect.

For example: #101173 describes a SymPy simplification bug that doesn't reach TV, since
it's run only in the end of the tracing.

ghstack-source-id: f0175c78284b42e0aaf91476f4be693d4fafb17c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106645