expandable_segments fix possible assert
If record_history is enabled, then a block is allocated, record_history
is disabled, and then the block is freed and later unnmapped, we can hit
the `to_map->context_when_allocated == nullptr` assertion.

This change universally clears context_when_allocated on free, which should
prevent this sequence of events from  happening.

ghstack-source-id: 8805568f97f1a14adbe7589ff9255152241c0fef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106818