Use expect_true to make split with unbacked sizes work.
This pattern shows up in torchrec KeyedJaggedTensor.  Most
of the change in this PR is mechanical: whenever we failed
an unbacked symint test due to just error checking, replace the
conditional with something that calls expect_true (e.g.,
torch._check or TORCH_SYM_CHECK).

Some of the changes are a bit more nuanced, I've commented on the PR
accordingly.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 11d5692b882a2825312335b046cf1c6d8e798e95
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106788